Fix keyboard events disruption issue
I resolved an issue where keyboard events were disrupted after executing
the `LoadingScreen.loadAssets` method. The problem was probably related
to the AWT event loop and occurred when the method was called after creating the game window.

To address this, I made the method static and ensured it is called before the game window creation.